pyproject.toml
[project]
name = "gendiff" 
version = "0.1.0"
description = "Compares two configuration files and shows a difference"
authors = [
    { name = "Targitay", email = "cibrenok@bk.ru" },
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0.3",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gendiff"]
